split partial magic system
