+++ Content.Server/_CP14/MagicEnergy/CP14MagicEnergySystem.Relay.cs
@@ -1,0 +1,6 @@
+namespace Content.Server._CP14.MagicEnergy;
+
+public partial class CP14MagicEnergySystem
+{
+
+}
