+++ Content.Server/_CP14/MagicEnergy/Components/CP14MagicEnergyRelayComponent.cs
@@ -1,0 +1,9 @@
+namespace Content.Server._CP14.MagicEnergy.Components;
+
+/// <summary>
+///
+/// </summary>
+[RegisterComponent, Access(typeof(CP14MagicEnergySystem))]
+public sealed partial class CP14MagicEnergyRelayComponent : Component
+{
+}
